// src/lib/api/types/createGuideTranslation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CreateGuideTranslation {
  /** Localized description */
  description?: string;
  /**
   * Language code (en, am)
   * @minLength 2
   * @maxLength 5
   */
  language: string;
  /**
   * Localized name
   * @minLength 1
   * @maxLength 200
   */
  name: string;
}