// src/lib/api/types/createStepTranslation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CreateStepTranslation {
  /** Localized description */
  description?: string;
  /** Rich content as JSON */
  detailedContent?: unknown;
  /**
   * Language code (en, am)
   * @minLength 2
   * @maxLength 5
   */
  language: string;
  /** Required documents as JSON array */
  requiredDocuments?: unknown;
  /**
   * Localized title
   * @minLength 1
   * @maxLength 200
   */
  title: string;
}